A First Kotlin Android app for self-learning
A very simple and ugly Android app, in Kotlin, in an attempt to teach myself
modern Android app development.

Very simply, this app reaches out to a free API to retrieve JSON
containing the information pertaining to a random Magic: the Gathering
card. It then extracts the flavor text for the card and displays it
in the app.